btrfs-progs: crypto: clean up types and includes for crc32c
There are some stale headers that we don't need and the int types are
using the kernel types and pull kerncompat.h. As this is a basic header
that should minimize dependencies use the standard int types.

+++ crypto/crc32c.h
@@ -19,9 +19,9 @@
 #ifndef __CRC32C__
 #define __CRC32C__
 
-#include "kerncompat.h"
+#include <inttypes.h>
 
-u32 crc32c_le(u32 seed, unsigned char const *data, size_t length);
+uint32_t crc32c_le(uint32_t seed, unsigned char const *data, uint32_t length);
 void crc32c_optimization_init(void);
 
 #define crc32c(seed, data, length) crc32c_le(seed, (unsigned char const *)data, length)
